+++ CHANGELOG.md
@@ -1,5 +1,56 @@
 # Change log
 
+## V. 0.1.0 - May 20. 2020
+
+**Added**
+-  Added Navigation
+-  Added 2.png
+-  Added 3.png
+-  Added 4.png
+-  Added 5.png
+-  Added 6.png
+-  Added 7.png
+-  Added 8.png
+-  Added 9.png  
+-  Added 10.png
+-  Added 11.png
+-  Added 12.png  
+-  Added 13.png
+-  Added 14.png
+-  Added 15.png  
+-  Added 16.png
+-  Added 17.png
+-  Added 18.png  
+
+**Changed**
+
+-  Changed images in Home Screen
+-  Finish Home Screen
+-  Custom Navigation
+
+## V. 0.0.4 - May 19. 2020
+
+**Added**
+-  Create Section Containers
+-  Create Gap component 
+
+**Fixed**
+- Fixed position components in Home Screen 
+
+## V. 0.0.3 - May 19. 2020
+
+**Added**
+-  Create Section Containers
+
+**Fixed**
+- Fixed position components in Home Screen 
+
+## V. 0.0.2 - May 18. 2020
+
+**Added**
+-  Create Section Containers
+
+
 ## V. 0.0.1 - May 18. 2020
 
 Initial relase.
@@ -11,11 +62,11 @@ Initial relase.
 **Added:**
 
 -   Change log
--   splash.png
--   icon.png
--   forest.png
--   1.png
--   Light Bulb
--   Home Title
+-   Added splash.png
+-   Added icon.png
+-   Added forest.png
+-   Added 1.png
+-   Added Light Bulb
+-   Create Home Title